Accept, with modification, numbers 1, 4, and 5 of 5 in SourceForge patch 3152896 from dmichelsen (Dagobert Michelsen): Make the first parameter of VotingBooth::vote() const to match its declaration. Cast integer logf() argument to float to avoid ambiguity. Declare HUDuiList::createSlider() parameter as const to match its implementation.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21900 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/ServerMenu.cxx
@@ -410,7 +410,7 @@ void ServerMenu::setSelected(int index, bool forcerefresh)
 	}
 	else {
 	  // colorize servers: many shots->red, jumping->green, CTF->blue
-	  const float rf = std::min(1.0f, logf(server.ping.maxShots) / logf(20.0f));
+	  const float rf = std::min(1.0f, logf((float)server.ping.maxShots) / logf(20.0f));
 	  const float gf = gameOptions & JumpingGameStyle ? 1.0f : 0.0f;
 	  const float bf = (gameType == ClassicCTF) ? 1.0f : 0.0f;
 	  label->setColor(0.5f + rf * 0.5f, 0.5f + gf * 0.5f, 0.5f + bf * 0.5f);
